+++ app/serializers/dc20/character_serializer.rb
@@ -10,7 +10,7 @@ module Dc20
                :language_points, :trades, :trade_knowledge, :language_levels, :attribute_saves, :physical_save, :mental_save,
                :initiative, :conditions, :attacks, :health, :path_points, :paths, :mana_points, :maneuvers, :maneuver_points,
                :stamina_points, :grit_points, :rest_points, :mana_spend_limit, :cantrips, :spells, :talent_points,
-               :spell_lists_amount, :selected_talents, :subclass, :speeds, :jump, :size, :breath
+               :spell_lists_amount, :selected_talents, :subclass, :speeds, :jump, :size, :breath, :spell_list
 
     delegate :features, :id, :name, :level, :main_class, :abilities, :modified_abilities, :health, :attribute_points, :classes,
              :ancestries, :combat_mastery, :save_dc, :precision_defense, :area_defense, :attack, :skills, :skill_points,
@@ -20,7 +20,7 @@ module Dc20
              :mana_spend_limit, :cantrips, :spells, :talent_points, :spell_lists_amount, :speeds, :jump, :size, :breath,
              to: :decorator
     delegate :created_at, :updated_at, :data, to: :object
-    delegate :guide_step, :selected_talents, :subclass, to: :data
+    delegate :guide_step, :selected_talents, :subclass, :spell_list, to: :data
 
     def provider
       'dc20'