+++ config/initializers/container.rb
@@ -47,6 +47,9 @@ module Charkeeper
     register('commands.characters_context.dc20.feats.add') { CharactersContext::Dc20::Feats::AddCommand.new }
     register('commands.characters_context.dc20.rest.perform') { CharactersContext::Dc20::Rest::PerformCommand.new }
 
+    register('commands.characters_context.dc20.spells.add') { CharactersContext::Dc20::Spells::AddCommand.new }
+    register('commands.characters_context.dc20.spells.change') { CharactersContext::Dc20::Spells::ChangeCommand.new }
+
     register('commands.characters_context.dnd5.create') { CharactersContext::Dnd5::CreateCommand.new }
     register('commands.characters_context.dnd5.update') { CharactersContext::Dnd5::UpdateCommand.new }
     register('commands.characters_context.dnd5.spell_update') { CharactersContext::Dnd5::SpellUpdateCommand.new }